requirements.txt
appdirs==1.4.4
fpdf2==2.4.6
Pillow==8.4.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.3
